add default sortObj for search. Cleanup
